Move DomainServiceEndpointRoutePatternAttribute to Hosting
